# swagger.yaml
openapi: 3.0.0
info:
  title: Express API Starter Kit
  version: 1.0.0
  description: API documentation for Express API starter kit
servers:
  - url: http://localhost:4000/v1
    description: HTTP API Starter Kit
  - url: https://sandbox.api.yourdomain.com/v1
    description: Sandbox environment
  - url: https://api.yourdomain.com/v1
    description: Production environment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  bearerAuth: []
tags:
  - name: Auth
    description: API endpoints for managing Authentication
  - name: Users
    description: API endpoints for managing users
paths:
  /register:
    post:
      summary: Register a new user
      description: Create a new user with email and password.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - confirmPassword
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: Pas$word123
                confirmPassword:
                  type: string
                  format: password
                  example: Pas$word123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 64e72f3bb6dc7b001f2a8c12
                  name:
                    type: string
                    example: John Doe
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR...
  /login:
    post:
      summary: Login user
      description: Authenticate user using email and password.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: Pas$word123
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 64e72f3bb6dc7b001f2a8c12
                      name:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: john@example.com
                      status:
                        type: string
                        example: active
        '401':
          description: Invalid email or password
  /users:
    get:
      summary: Get all users
      description: Retrieve a list of users, optionally filtered.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - active
              - inactive
          description: Filter users by status
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of results per page
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fulfilled
                  message:
                    type: string
                    example: Users list
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 64e72f3bb6dc7b001f2a8c12
                        name:
                          type: string
                          example: John Doe
                        email:
                          type: string
                          example: john@example.com
                        status:
                          type: string
                          example: active
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Server error
    post:
      summary: Create a new user
      description: Create a user account with name, email, and status.
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - status
              properties:
                name:
                  type: string
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                status:
                  type: string
                  enum:
                    - active
                    - inactive
                  example: active
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fulfilled
                  message:
                    type: string
                    example: User created successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 64e72f3bb6dc7b001f2a8c12
                      name:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: john@example.com
                      status:
                        type: string
                        example: active
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Invalid input
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
  /users/{id}:
    get:
      summary: Get a user by ID
      description: Retrieve a single user by their unique ID.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fulfilled
                  message:
                    type: string
                    example: User fetched successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 64e72f3bb6dc7b001f2a8c12
                      name:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: john@example.com
                      status:
                        type: string
                        example: active
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
    put:
      summary: Update a user
      description: Update a user's name, email, and status by ID.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - status
              properties:
                name:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: jane@example.com
                status:
                  type: string
                  enum:
                    - active
                    - inactive
                  example: inactive
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fulfilled
                  message:
                    type: string
                    example: User updated successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 64e72f3bb6dc7b001f2a8c12
                      name:
                        type: string
                        example: Jane Doe
                      email:
                        type: string
                        example: jane@example.com
                      status:
                        type: string
                        example: inactive
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Invalid input
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
    delete:
      summary: Delete a user
      description: Delete a user by their unique ID.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to delete
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fulfilled
                  message:
                    type: string
                    example: User deleted successfully
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Invalid user ID
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: rejected
                  message:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error